Created IsolateChannel and messages to send information between isolates

+++ bin/arch_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 import 'dart:isolate';
 
+import 'package:arch_test/isolate_channel.dart';
 import 'package:path/path.dart';
 
 void main() async {
@@ -24,14 +25,18 @@ void main() async {
     flush: true,
   );
 
-  final receivePort = ReceivePort();
+  final channel = IsolateChannel();
 
-  receivePort.listen((message) async {
-    receivePort.close();
-    await archTestRunnerFile.delete();
+  channel.messages.listen((message) async {
+    if (message is PrintIsolateMessage) {
+      print(message.message);
+    } else if (message is ProcessFinishedIsolateMessage) {
+      await archTestRunnerFile.delete();
+      channel.close();
+    }
   });
 
-  await Isolate.spawnUri(archTestRunnerFile.uri, [], receivePort.sendPort);
+  await Isolate.spawnUri(archTestRunnerFile.uri, [], channel.sendPort);
 }
 
 const _archTestRunnerFileContent = '''import 'dart:io';
@@ -39,18 +44,31 @@ import 'dart:isolate';
 
 import 'package:analyzer/dart/element/element.dart';
 import 'package:arch_test/arch_test.dart';
+import 'package:arch_test/isolate_channel.dart';
 
 import 'arch_test.dart' as tests;
 
 void main(List<String> args, SendPort sendPort) async {
+  final channel = IsolateChannel.fromSendPort(sendPort);
+
+  channel.send(PrintIsolateMessage(message: 'Loading package libraries...'));
+
   final packageLoader = PackageLoader();
   final packageLibraries = await packageLoader.loadLibraries(Directory.current);
 
+  channel.send(PrintIsolateMessage(message: 'Package libraries loaded!'));
+  channel.send(PrintIsolateMessage(message: 'Registering architecture rules...'));
+
   tests.main();
+
+  channel.send(PrintIsolateMessage(message: 'Architecture rules registered!'));
+  channel.send(PrintIsolateMessage(message: 'Checking architecture rules...'));
 
   await _checkRules(packageLibraries);
 
-  sendPort.send(null);
+  channel.send(PrintIsolateMessage(message: 'Architecture rules checked!'));
+
+  channel.send(const ProcessFinishedIsolateMessage());
 }
 
 Future<void> _checkRules(List<LibraryElement> packageLibraries) async {
